requirements.txt
Flask
Flask-Session
Flask-Cors
mysql-connector-python
python-dotenv
Werkzeug